+++ src/Control/Concurrent/NQE.hs
@@ -2,8 +2,10 @@ module Control.Concurrent.NQE
     ( module Control.Concurrent.NQE.Process
     , module Control.Concurrent.NQE.Network
     , module Control.Concurrent.NQE.Supervisor
+    , module Control.Concurrent.NQE.PubSub
     ) where
 
 import           Control.Concurrent.NQE.Network
 import           Control.Concurrent.NQE.Process
+import           Control.Concurrent.NQE.PubSub
 import           Control.Concurrent.NQE.Supervisor