+++ CHANGELOG.md
@@ -12,3 +12,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Type-safe asynchronous messages.
 - Supervisors for `MonadUnliftIO` actions.
 - Test suite.
+- PubSub actor.